Improved map height
Increase the height difference in the map with a height multiplyer, and added a curve to control the change in height depending on the noise level. This keeps all of the noise below .4 at 0 so that the water appears level, and allows us to change the curvature of the mountains.